+++ tests/Mailtrap.IntegrationTests/TestConstants/UrlSegmentsTestConstants.cs
@@ -27,4 +27,5 @@ internal static class UrlSegmentsTestConstants
     internal static string EventsSegment { get; } = "events";
     internal static string BatchEmailSegment { get; } = "batch";
     internal static string StatsSegment { get; } = "stats";
+    internal static string ApiTokensSegment { get; } = "api_tokens";
 }